chg: move the base for multistate widgets inside the widget module, use it for other widgets
